third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/var-evaluation-expected.txt

This is a testharness.js-based test.
[FAIL] inline-size with var-substitution
  assert_equals: expected "rgb(0, 128, 0)" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
